chore: adopt a committed golangci config, and fix the ~350 issues rather than excluding them (#19)
Adopts the file-service .golangci.yml. It reported 1073 issues against this
repository. The config now reports ZERO with no exclusions section and no nolint
directives anywhere in the tree — the difference was closed by changing code, not
configuration, because a rule excluded everywhere is a rule you have not adopted.

Four linters from the source config are NOT enabled. Each is recorded in
docs/LIBRARY-GAPS.md with its count and the reason, rather than silenced here:
revive's `exported` (239 — the real gap, missing godoc), gosec (227), revive's
`unused-parameter` (131), gocyclo (113), and revive's `bool-literal-in-expr` (11,
where 11 of 13 compare an interface{} to a bool literal and the suggested
rewrite does not compile).

gosec was sampled by hand rather than dismissed: G115 at readV2Column is guarded
by `size > uint64(decoder.Len())`, and G602 at abstract.go by `if left == 0`.
Its findings here are dominated by guards its flow analysis cannot see, and
verifying all 145 G115 sites is the same manual trace the panic audit required.

BREAKING — public naming brought in line with Go conventions:

  Doc.Guid            -> Doc.GUID
  Doc.GetSubdocGuids  -> Doc.GetSubdocGUIDs
  Doc.GetXmlFragment  -> Doc.GetXMLFragment
  ToJson              -> ToJSON   on Doc, YArray, YMap, YText, YXmlFragment

THE RENAME EXPOSED A REAL API WART. YXmlText carried BOTH an inherited
ToJson() interface{} and its own ToJSON() string, returning DIFFERENT strings:
YText.ToString() is plain text, YXmlText.ToString() builds the XML form. Two
methods differing only in capitalisation, and nothing pointed at it while the
names differed. The public method now keeps each type's Yjs-correct signature and
the polymorphism moved to an unexported toJSONValue().

THE LINTERS CAUGHT A BUG I INTRODUCED, which is the most useful thing here.
Rewriting if/else to switch for gocritic's ifElseChain silently changes `break`
semantics: the break that exited the LOOP starts exiting the SWITCH. In merge.go
that turned "we are done" into an infinite loop; govet and staticcheck flagged it
immediately. item.go's integration loop had THREE loop-breaking breaks inside a
chain gocritic wanted rewritten. All now use labelled breaks, and
docs/LIBRARY-GAPS.md warns not to run --fix on this repository without reading
the diff.

Disabling revive's `exported` also does not make that gap go away: staticcheck
enforces the same thing via ST1020/21/22 and golangci-lint had been
deduplicating them. The ~70 MALFORMED comments it flags are fixed here; the 239
missing ones remain the recorded gap.

Documentation defects found by fixing those:

  - A 25-line block documenting decodeAwarenessEntries sat above
    ParseAwarenessStateJSON. Moved.
  - YArray.Splice and YXmlFragment.Slice both carried ToArray's description.
  - Section banners were being absorbed into the following doc comment.

Also: protocol.InspectMessage wrapped an inner error with %v instead of %w so it
could not be unwrapped; unconvert removed 197 no-op conversions (Number is an
alias for int, so they were free but noise); several internal signatures lost
parameters and returns no caller used.

New: SECURITY.md, CONTRIBUTING.md, CHANGELOG.md. SECURITY.md states the actual
threat model — this library decodes untrusted bytes, and anything reachable from
the decoders that panics or hangs is a vulnerability here, not caller error —
and records that the panic audit was a trace through the guards rather than "the
fuzzer has not found one".

Verified: gofmt clean, vet clean in BOTH build configurations (the tagged one
caught a file the default vet cannot see), golangci-lint 0 issues, full suite and
whole-module -race green, and the oracle gate passed at 13 surfaces both
directions 20k seeds after every core-code change.

One self-inflicted failure worth recording: the rename initially hit
bench/ygo/main.go, a separate module benchmarking the ygo library, which has its
own GetXmlFragment. `go build ./...` could not see it because it is a different
module; the bracket-protocol test did. Reverted.

+++ crdt/content_string_test.go
@@ -451,7 +451,7 @@ func TestContentStringUTF16IndexThresholdAndSplitPreservation(t *testing.T) {
 		t.Fatal("short non-ASCII string retained a sampled index")
 	}
 
-	const runes = int(contentStringUTF16IndexThreshold * 3)
+	const runes = contentStringUTF16IndexThreshold * 3
 	content := newContentString(strings.Repeat("界", runes))
 	length := content.contentLength()
 	splitAt := contentStringUTF16IndexThreshold + 32
@@ -460,10 +460,10 @@ func TestContentStringUTF16IndexThresholdAndSplitPreservation(t *testing.T) {
 	if index == nil || right.utf16Index != index {
 		t.Fatalf("long split did not share its sampled index: left=%p right=%p", index, right.utf16Index)
 	}
-	if got, want := content.value, strings.Repeat("界", int(splitAt)); got != want {
+	if got, want := content.value, strings.Repeat("界", splitAt); got != want {
 		t.Fatalf("indexed left = %q, want %q", got, want)
 	}
-	if got, want := right.value, strings.Repeat("界", runes-int(splitAt)); got != want {
+	if got, want := right.value, strings.Repeat("界", runes-splitAt); got != want {
 		t.Fatalf("indexed right = %q, want %q", got, want)
 	}
 	if got := right.contentLength(); got != length-splitAt {
